backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
azure-cosmos==4.5.0
azure-storage-blob==12.19.1
openai==1.30.1
stripe==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymupdf==1.24.2
python-docx==1.1.0
cryptography==42.0.5
pydantic-settings==2.2.1
